apps/backend/pyproject.toml
[project]
name = "ai-email-assistant-backend"
version = "0.1.0"
description = "Backend for AI Email Assistant"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
  "flask>=3.0.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.1",
  "SQLAlchemy>=2.0.0",
  "psycopg2-binary>=2.9.9",
  "apscheduler>=3.10.4",
  "langgraph>=0.2.0",
  "google-generativeai>=0.7.0",
  "playwright>=1.44.0",
  "pyjwt>=2.8.0",
  "requests>=2.31.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
